pyproject.toml
[tool.poetry]
name = "taipy-demo-job-monitoring"
version = "0.1.0"
description = ""
authors = ["Grégoire Marabout-Demazure <gmarabout@gmail.com>"]
readme = "README.md"
packages = [{ include = "job_monitoring" }]

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.1.3"
numpy = "^1.23.5"
taipy-core = "^2.0.4"
taipy-gui = "^2.0.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = { version = "^22.10.0", allow-prereleases = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
